refactor: Move wasm implementation to separate files
